+++ hydromt_wflow/utils.py
@@ -60,7 +60,7 @@ of the config.
             usecols = np.append(usecols, np.arange(count, count + len(gdf.index)))
             count += len(gdf.index)
             da_ts = open_timeseries_from_table(
-                fn, name=f'{header}_{col["map"]}', usecols=usecols
+                fn, name=f"{header}_{col['map']}", usecols=usecols
             )
             da = GeoDataArray.from_gdf(gdf, da_ts, index_dim="index")
         # Column based on xy coordinates or reducer for the full model domain domain
@@ -105,7 +105,7 @@ of the config.
                     # Dimensions are ascending and ordered as (x,y,layer,time)
                     # Indices are created before ordering for compatibility with
                     # raster.idx_to_xy
-                    full_index = maps[f'{config["input"].get("subcatchment")}'].copy()
+                    full_index = maps[f"{config['input'].get('subcatchment')}"].copy()
                     res_x, res_y = full_index.raster.res
                     if res_y < 0:
                         full_index = full_index.reindex(
@@ -128,18 +128,18 @@ of the config.
                         "vertical" in col["parameter"]
                         or "lateral.land" in col["parameter"]
                     ):
-                        mask = maps[f'{config["input"].get("subcatchment")}'].copy()
+                        mask = maps[f"{config['input'].get('subcatchment')}"].copy()
                     elif "reservoir" in col["parameter"]:
                         mask = maps[
-                            f'{config["input"]["lateral"]["river"]["reservoir"].get("locs")}'
+                            f"{config['input']['lateral']['river']['reservoir'].get('locs')}"
                         ].copy()
                     elif "lake" in col["parameter"]:
                         mask = maps[
-                            f'{config["input"]["lateral"]["river"]["lake"].get("locs")}'
+                            f"{config['input']['lateral']['river']['lake'].get('locs')}"
                         ].copy()
                     # Else lateral.river
                     else:
-                        mask = maps[f'{config["input"].get("river_location")}'].copy()
+                        mask = maps[f"{config['input'].get('river_location')}"].copy()
                     # Rearrange the mask
                     res_x, res_y = mask.raster.res
                     if res_y < 0: